+++ src/main/java/com/netj/deungchi/domain/Mountain.java
@@ -32,8 +32,8 @@ public class Mountain {
     private String level;
     private String average_time;
     private String altitude;
-    private Float latitude;
-    private Float longitude;
+    private Double latitude;
+    private Double longitude;
     private String featured_image;
 
     @CreationTimestamp
