+++ src/main/java/com/netj/deungchi/domain/Record.java
@@ -50,7 +50,7 @@ public class Record {
     String hikingDuration;
 
     // 등산 거리
-    Float hikingLength;
+    Double hikingLength;
 
     Integer temperature;
 
@@ -63,7 +63,7 @@ public class Record {
     Timestamp updatedAt;
 
     @Builder
-    public Record(Member member, Mountain mountain, Course course, CourseDetail courseDetail,String level, String content, Timestamp startAt, Timestamp endAt, String startLocation, String endLocation, String hikingDuration, Float hikingLength, Integer temperature, Boolean isShare) {
+    public Record(Member member, Mountain mountain, Course course, CourseDetail courseDetail,String level, String content, Timestamp startAt, Timestamp endAt, String startLocation, String endLocation, String hikingDuration, Double hikingLength, Integer temperature, Boolean isShare) {
         this.member = member;
         this.mountain = mountain;
         this.course = course;
